+++ tests/compile_fail/invalid_size_attr.rs
@@ -4,12 +4,12 @@ fn main() {}
 
 #[derive(ShaderType)]
 struct Test {
-    #[size]
+    #[shader(size)]
     a: u32,
-    #[size()]
+    #[shader(size())]
     b: u32,
-    #[size(invalid)]
+    #[shader(size(invalid))]
     c: u32,
-    #[size(-1)]
+    #[shader(size(-1))]
     d: u32,
 }